Move name utilities from StringUtils to NameUtils
