+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/internal/DataflowUnboundedReadFromBoundedSource.java
@@ -19,7 +19,6 @@ package org.apache.beam.runners.dataflow.internal;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static org.apache.beam.sdk.util.StringUtils.approximateSimpleName;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -50,6 +49,7 @@ import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.transforms.PTransform;
 import org.apache.beam.sdk.transforms.display.DisplayData;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
+import org.apache.beam.sdk.util.NameUtils;
 import org.apache.beam.sdk.util.PropertyNames;
 import org.apache.beam.sdk.values.PBegin;
 import org.apache.beam.sdk.values.PCollection;
@@ -109,10 +109,10 @@ public class DataflowUnboundedReadFromBoundedSource<T> extends PTransform<PBegin
     if (source.getClass().isAnonymousClass()) {
       sourceName = "AnonymousSource";
     } else {
-      sourceName = approximateSimpleName(source.getClass());
+      sourceName = NameUtils.approximateSimpleName(source.getClass());
     }
 
-    return "Read(" + sourceName + ")";
+    return String.format("Read(%s)", sourceName);
   }
 
   @Override
