+++ sdks/java/core/src/test/java/org/apache/beam/sdk/util/StringUtilsTest.java
@@ -20,13 +20,7 @@ package org.apache.beam.sdk.util;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
-import org.apache.beam.sdk.io.TextIO;
-import org.apache.beam.sdk.transforms.PTransform;
-import org.apache.beam.sdk.values.PBegin;
-import org.apache.beam.sdk.values.PDone;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
@@ -35,8 +29,6 @@ import org.junit.runners.JUnit4;
  */
 @RunWith(JUnit4.class)
 public class StringUtilsTest {
-  @Rule
-  public ExpectedException thrown = ExpectedException.none();
 
   @Test
   public void testTranscodeEmptyByteArray() {
@@ -53,98 +45,6 @@ public class StringUtilsTest {
     String string = "%00%05%0c%10%1f AB~%7f%80%ffC%00";
     assertEquals(string, StringUtils.byteArrayToJsonString(bytes));
     assertArrayEquals(bytes, StringUtils.jsonStringToByteArray(string));
-  }
-
-  /**
-   * Inner class for simple name test.
-   */
-  private class EmbeddedOldDoFn {
-
-    private class DeeperEmbeddedOldDoFn extends EmbeddedOldDoFn {}
-
-    private EmbeddedOldDoFn getEmbedded() {
-      return new DeeperEmbeddedOldDoFn();
-    }
-  }
-
-  private class EmbeddedPTransform extends PTransform<PBegin, PDone> {
-    @Override
-    public PDone expand(PBegin begin) {
-      throw new IllegalArgumentException("Should never be applied");
-    }
-
-    private class Bound extends PTransform<PBegin, PDone> {
-      @Override
-      public PDone expand(PBegin begin) {
-        throw new IllegalArgumentException("Should never be applied");
-      }
-    }
-
-    private Bound getBound() {
-      return new Bound();
-    }
-  }
-
-  private interface AnonymousClass {
-    Object getInnerClassInstance();
-  }
-
-  @Test
-  public void testSimpleName() {
-    assertEquals("Embedded",
-        StringUtils.approximateSimpleName(EmbeddedOldDoFn.class));
-  }
-
-  @Test
-  public void testAnonSimpleName() throws Exception {
-    thrown.expect(IllegalArgumentException.class);
-
-    EmbeddedOldDoFn anon = new EmbeddedOldDoFn(){};
-
-    StringUtils.approximateSimpleName(anon.getClass());
-  }
-
-  @Test
-  public void testNestedSimpleName() {
-    EmbeddedOldDoFn fn = new EmbeddedOldDoFn();
-    EmbeddedOldDoFn inner = fn.getEmbedded();
-
-    assertEquals("DeeperEmbedded", StringUtils.approximateSimpleName(inner.getClass()));
-  }
-
-  @Test
-  public void testPTransformName() {
-    EmbeddedPTransform transform = new EmbeddedPTransform();
-    assertEquals(
-        "StringUtilsTest.EmbeddedPTransform",
-        StringUtils.approximatePTransformName(transform.getClass()));
-    assertEquals(
-        "StringUtilsTest.EmbeddedPTransform",
-        StringUtils.approximatePTransformName(transform.getBound().getClass()));
-    assertEquals("TextIO.Write", StringUtils.approximatePTransformName(TextIO.Write.Bound.class));
-  }
-
-  @Test
-  public void testPTransformNameWithAnonOuterClass() throws Exception {
-    AnonymousClass anonymousClassObj = new AnonymousClass() {
-      class NamedInnerClass extends PTransform<PBegin, PDone> {
-        @Override
-        public PDone expand(PBegin begin) {
-          throw new IllegalArgumentException("Should never be applied");
-        }
-      }
-
-      @Override
-      public Object getInnerClassInstance() {
-        return new NamedInnerClass();
-      }
-    };
-
-    assertEquals("NamedInnerClass",
-        StringUtils.approximateSimpleName(anonymousClassObj.getInnerClassInstance().getClass()));
-    assertEquals("StringUtilsTest.NamedInnerClass",
-        StringUtils.approximatePTransformName(
-            anonymousClassObj.getInnerClassInstance().getClass()));
   }
 
   @Test
